Split into QueryableInputObjectTypeResolver
