+++ layers/Engine/packages/component-model/src/FieldResolvers/ObjectType/AbstractQueryableObjectTypeFieldResolver.php
@@ -9,7 +9,7 @@ use PoP\ComponentModel\ModuleProcessors\FilterInputContainerModuleProcessorInter
 use PoP\ComponentModel\ModuleProcessors\ModuleProcessorManagerInterface;
 use PoP\ComponentModel\Resolvers\QueryableFieldResolverTrait;
 use PoP\ComponentModel\Resolvers\QueryableInterfaceSchemaDefinitionResolverAdapter;
-use PoP\ComponentModel\TypeResolvers\InputObjectType\InputObjectTypeResolverInterface;
+use PoP\ComponentModel\TypeResolvers\InputObjectType\QueryableInputObjectTypeResolverInterface;
 use PoP\ComponentModel\TypeResolvers\ObjectType\ObjectTypeResolverInterface;
 
 abstract class AbstractQueryableObjectTypeFieldResolver extends AbstractObjectTypeFieldResolver implements QueryableObjectTypeFieldSchemaDefinitionResolverInterface
@@ -116,7 +116,7 @@ abstract class AbstractQueryableObjectTypeFieldResolver extends AbstractObjectTy
         $consolidatedFieldArgNameTypeResolvers = $this->getConsolidatedFieldArgNameTypeResolvers($objectTypeResolver, $fieldName);
         foreach ($fieldArgs as $fieldArgName => $fieldArgValue) {
             $fieldArgTypeResolver = $consolidatedFieldArgNameTypeResolvers[$fieldArgName];
-            if (!($fieldArgTypeResolver instanceof InputObjectTypeResolverInterface)) {
+            if (!($fieldArgTypeResolver instanceof QueryableInputObjectTypeResolverInterface)) {
                 continue;
             }
             $fieldArgTypeResolver->integrateInputValueToFilteringQueryArgs($filteringQueryArgs, $fieldArgValue);
