+++ layers/Engine/packages/component-model/src/TypeResolvers/InputObjectType/AbstractInputObjectTypeResolver.php
@@ -7,8 +7,6 @@ namespace PoP\ComponentModel\TypeResolvers\InputObjectType;
 use Exception;
 use PoP\ComponentModel\ComponentConfiguration;
 use PoP\ComponentModel\ErrorHandling\Error;
-use PoP\ComponentModel\FilterInputProcessors\FilterInputProcessorInterface;
-use PoP\ComponentModel\FilterInputProcessors\FilterInputProcessorManagerInterface;
 use PoP\ComponentModel\Resolvers\TypeSchemaDefinitionResolverTrait;
 use PoP\ComponentModel\Schema\InputCoercingServiceInterface;
 use PoP\ComponentModel\Schema\SchemaTypeModifiers;
@@ -34,7 +32,6 @@ abstract class AbstractInputObjectTypeResolver extends AbstractTypeResolver impl
 
     private ?DangerouslyDynamicScalarTypeResolver $dangerouslyDynamicScalarTypeResolver = null;
     private ?InputCoercingServiceInterface $inputCoercingService = null;
-    private ?FilterInputProcessorManagerInterface $filterInputProcessorManager = null;
 
     final public function setDangerouslyDynamicScalarTypeResolver(DangerouslyDynamicScalarTypeResolver $dangerouslyDynamicScalarTypeResolver): void
     {
@@ -51,14 +48,6 @@ abstract class AbstractInputObjectTypeResolver extends AbstractTypeResolver impl
     final protected function getInputCoercingService(): InputCoercingServiceInterface
     {
         return $this->inputCoercingService ??= $this->instanceManager->getInstance(InputCoercingServiceInterface::class);
-    }
-    final public function setFilterInputProcessorManager(FilterInputProcessorManagerInterface $filterInputProcessorManager): void
-    {
-        $this->filterInputProcessorManager = $filterInputProcessorManager;
-    }
-    final protected function getFilterInputProcessorManager(): FilterInputProcessorManagerInterface
-    {
-        return $this->filterInputProcessorManager ??= $this->instanceManager->getInstance(FilterInputProcessorManagerInterface::class);
     }
 
     public function getInputFieldDescription(string $inputFieldName): ?string
@@ -405,48 +394,5 @@ abstract class AbstractInputObjectTypeResolver extends AbstractTypeResolver impl
     protected function validateInputFieldValue(string $inputFieldName, mixed $inputFieldValue): array
     {
         return [];
-    }
-    public function getInputFieldFilterInput(string $inputFieldName): ?array
-    {
-        return null;
-    }
-    /**
-     * If the input field is an InputObject, then forward the logic.
-     * Otherwise, if the input field defines a FilterInput,
-     * apply it to obtain the filtering query
-     *
-     * @param array<string, mixed> $query
-     */
-    final public function integrateInputValueToFilteringQueryArgs(array &$query, stdClass $inputValue): void
-    {
-        foreach ((array)$inputValue as $inputFieldName => $inputFieldValue) {
-            $this->integrateInputFieldValueToFilteringQueryArgs($inputFieldName, $query, $inputFieldValue);
-        }
-    }
-    /**
-     * Integrate an InputObject into the filtering args.
-     * 
-     * By default, forward the logic to its contained input fields
-     */
-    protected function integrateInputFieldValueToFilteringQueryArgs(string $inputFieldName, array &$query, mixed $inputFieldValue): void
-    {
-        /**
-         * If the input field defines a FilterInput, apply it to obtain the filtering query
-         */
-        if ($filterInput = $this->getInputFieldFilterInput($inputFieldName)) {
-            /** @var FilterInputProcessorInterface */
-            $filterInputProcessor = $this->getFilterInputProcessorManager()->getProcessor($filterInput);
-            $filterInputProcessor->filterDataloadQueryArgs($filterInput, $query, $inputFieldValue);
-            return;
-        }
-        
-        /**
-         * If the input field is an InputObject, recursively apply this function
-         */
-        $inputFieldNameTypeResolvers = $this->getConsolidatedInputFieldNameTypeResolvers();
-        $inputFieldTypeResolver = $inputFieldNameTypeResolvers[$inputFieldName];
-        if ($inputFieldTypeResolver instanceof InputObjectTypeResolverInterface) {
-            $inputFieldTypeResolver->integrateInputValueToFilteringQueryArgs($query, $inputFieldValue);
-        }
     }
 }
