+++ layers/Engine/packages/component-model/src/TypeResolvers/InputObjectType/InputObjectTypeResolverInterface.php
@@ -47,16 +47,4 @@ interface InputObjectTypeResolverInterface extends InputTypeResolverInterface
      * @return string[] Error messages
      */
     public function validateInputValue(stdClass $inputValue): array;
-    /**
-     * Have the input field filter the query
-     *
-     * @return array|null A FilterInput module, or null
-     */
-    public function getInputFieldFilterInput(string $inputFieldName): ?array;
-    /**
-     * Apply the FilterInputs to produce the filtering query
-     *
-     * @param array<string, mixed> $query
-     */
-    public function integrateInputValueToFilteringQueryArgs(array &$query, stdClass $inputValue): void;
 }
