+++ layers/Engine/packages/component-model/src/TypeResolvers/InputObjectType/QueryableInputObjectTypeResolverInterface.php
@@ -1,0 +1,28 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PoP\ComponentModel\TypeResolvers\InputObjectType;
+
+use stdClass;
+
+/**
+ * Based on GraphQL InputObject Type
+ *
+ * @see https://spec.graphql.org/draft/#sec-Input-Objects
+ */
+interface QueryableInputObjectTypeResolverInterface extends InputObjectTypeResolverInterface
+{
+    /**
+     * Have the input field filter the query
+     *
+     * @return array|null A FilterInput module, or null
+     */
+    public function getInputFieldFilterInput(string $inputFieldName): ?array;
+    /**
+     * Apply the FilterInputs to produce the filtering query
+     *
+     * @param array<string, mixed> $query
+     */
+    public function integrateInputValueToFilteringQueryArgs(array &$query, stdClass $inputValue): void;
+}
